[procon test] Run test src/atcoder/abc133/c.cc

+++ src/atcoder/abc133/c.cc
@@ -822,7 +822,7 @@ void body() {
     i64 ans = 2019;
     EACH(i, Ms) {
         EACH(j, Ms) {
-            ans = std::min(ans, (i * j) % 2019L);
+            ans = std::min(ans, i64((i * j) % 2019));
         }
     }
